Entity deserialization definitions
