+++ src/compiler/deserialization/mod.ts
@@ -1,5 +1,7 @@
 export * from './default.ucd-defs.js';
 export * from './ucd-def.js';
+export * from './ucd-entity-def.js';
+export * from './ucd-entity-prefix-def.js';
 export * from './ucd-function.js';
 export * from './ucd-lib.js';
 export * from './ucd-type-def.js';
