+++ src/compiler/deserialization/ucd-lib.spec.ts
@@ -22,7 +22,7 @@ describe('UcdLib', () => {
 
       expect(code).toContain("import('@hatsy/churi/deserializer')");
       expect(code).toContain('async readValue(stream, options) {\n');
-      expect(code).toMatch(/\bUcdReader\b/);
+      expect(code).toMatch(/\bAsyncUcdReader\b/);
       expect(code).not.toMatch(/\bcreateSyncUcdReader\b/);
     });
     it('creates sync factory', () => {
@@ -35,7 +35,7 @@ describe('UcdLib', () => {
       expect(code).toContain("import('@hatsy/churi/deserializer')");
       expect(code).toContain('readValue(input, options) {\n');
       expect(code).toMatch(/\bcreateSyncUcdReader\b/);
-      expect(code).not.toMatch(/\bUcdReader\b/);
+      expect(code).not.toMatch(/\bAsyncUcdReader\b/);
     });
     it('creates hybrid factory', () => {
       const compiled = lib.compile();
@@ -47,7 +47,7 @@ describe('UcdLib', () => {
       expect(code).toContain("import('@hatsy/churi/deserializer')");
       expect(code).toContain('readValue(input, options) {\n');
       expect(code).toMatch(/\bcreateSyncUcdReader\b/);
-      expect(code).toMatch(/\bUcdReader\b/);
+      expect(code).toMatch(/\bAsyncUcdReader\b/);
     });
   });
 
@@ -62,7 +62,7 @@ describe('UcdLib', () => {
       expect(new UccCode().write(module).toString()).toBe(code);
       expect(code).toContain(`} from '@hatsy/churi/deserializer';\n`);
       expect(code).toContain('export async function readValue(stream, options) {\n');
-      expect(code).toMatch(/\bUcdReader\b/);
+      expect(code).toMatch(/\bAsyncUcdReader\b/);
       expect(code).not.toMatch(/\bSyncUcdReader\b/);
     });
     it('compiles sync module', () => {
@@ -76,7 +76,7 @@ describe('UcdLib', () => {
       expect(code).toContain(`} from '@hatsy/churi/deserializer';\n`);
       expect(code).toContain('export function readValue(input, options) {\n');
       expect(code).toMatch(/\bcreateSyncUcdReader\b/);
-      expect(code).not.toMatch(/\bUcdReader\b/);
+      expect(code).not.toMatch(/\bAsyncUcdReader\b/);
     });
     it('compiles hybrid module', () => {
       const module = lib.compileModule();
@@ -89,7 +89,7 @@ describe('UcdLib', () => {
       expect(code).toContain(`} from '@hatsy/churi/deserializer';\n`);
       expect(code).toContain('export function readValue(input, options) {\n');
       expect(code).toMatch(/\bcreateSyncUcdReader\b/);
-      expect(code).toMatch(/\bUcdReader\b/);
+      expect(code).toMatch(/\bAsyncUcdReader\b/);
     });
   });
 });
