+++ src/deserializer/mod.ts
@@ -1,7 +1,7 @@
 /**
  * @module @hatsy/churi/deserializer
  */
-export * from './abstract-ucd-reader.js';
+export * from './async-ucd-reader.js';
 export * from './read-uc-list.js';
 export * from './sync-ucd-reader.js';
 export * from './ucd-reader.js';
