+++ src/deserializer/sync-ucd-reader.ts
@@ -1,11 +1,11 @@
 import { UcDeserializer } from '../schema/uc-deserializer.js';
 import { UcToken } from '../syntax/uc-token.js';
 import { UcTokenizer } from '../syntax/uc-tokenizer.js';
-import { AbstractUcdReader } from './abstract-ucd-reader.js';
 import { ucdReadValueSync } from './impl/ucd-read-value.sync.js';
+import { UcdReader } from './ucd-reader.js';
 import { UcdRx } from './ucd-rx.js';
 
-export class SyncUcdReader extends AbstractUcdReader {
+export class SyncUcdReader extends UcdReader {
 
   readonly #tokens: readonly UcToken[];
   #current = -1;
